+++ systems/renderer/helper.js
@@ -134,7 +134,7 @@ const getDirectionalLight = (directionalLight) => {
       [-prismRadius, 0, 0], [-prismRadius, 0, intensity],
       [0, prismRadius, 0], [0, prismRadius, intensity],
       [0, -prismRadius, 0], [0, -prismRadius, intensity]
-    ]
+    ],
   );
 };
 
@@ -151,7 +151,7 @@ const getPointLight = (pointLight) => {
       [0, -prismRadius, 0], [0, -radius, 0],
       [0, 0, prismRadius], [0, 0, radius],
       [0, 0, -prismRadius], [0, 0, -radius],
-    ]
+    ],
   );
 };
 
@@ -171,21 +171,21 @@ const getSpotLight = (spotLight) => {
       getPyramidEdgePositions({
         sx: radius * Math.sin(Math.PI / 4),
         sz: distance,
-      })
+      }),
     )
     .concat(
       getCirclePositions({
         radius,
         center: [0, 0, distance],
         ...spotLightCircleOptions,
-      })
+      }),
     )
     .concat(
       getCirclePositions({
         radius: innerRadius,
         center: [0, 0, distance],
         ...spotLightCircleOptions,
-      })
+      }),
     );
 };
 
@@ -325,7 +325,7 @@ export default ({ ctx }) => {
       const addToBuilder = (
         positions,
         color = [0.23, 0.23, 0.23, 1],
-        modelMatrix
+        modelMatrix,
       ) => {
         for (let i = 0; i < positions.length; i++) {
           const position = positions[i];
@@ -341,7 +341,7 @@ export default ({ ctx }) => {
         if (entity.transform?.position && entity.boundingBoxHelper) {
           addToBuilder(
             getBBoxPositionsList(entity.transform.worldBounds),
-            entity.boundingBoxHelper?.color || [1, 0, 0, 1]
+            entity.boundingBoxHelper?.color || [1, 0, 0, 1],
           );
         }
 
@@ -351,28 +351,28 @@ export default ({ ctx }) => {
             addToBuilder(
               getDirectionalLight(entity.directionalLight),
               entity.directionalLight.color,
-              modelMatrix
+              modelMatrix,
             );
           }
           if (entity.pointLight) {
             addToBuilder(
               getPointLight(entity.pointLight),
               entity.pointLight.color,
-              modelMatrix
+              modelMatrix,
             );
           }
           if (entity.spotLight) {
             addToBuilder(
               getSpotLight(entity.spotLight),
               entity.spotLight.color,
-              modelMatrix
+              modelMatrix,
             );
           }
           if (entity.areaLight) {
             addToBuilder(
               getAreaLight(entity.areaLight),
               entity.areaLight.color,
-              modelMatrix
+              modelMatrix,
             );
           }
         }
@@ -386,21 +386,21 @@ export default ({ ctx }) => {
               ? getOrthographicCamera(entity.camera)
               : getPerspectiveCamera(entity.camera),
             entity.cameraHelper.color,
-            modelMatrix
+            modelMatrix,
           );
         }
         if (entity.axesHelper) {
           addToBuilder(
             AXES_POSITIONS.map((p) => [...p]),
             AXES_COLORS.map((p) => [...p]),
-            modelMatrix
+            modelMatrix,
           );
         }
         if (entity.gridHelper) {
           addToBuilder(
             getGrid(entity.gridHelper),
             entity.gridHelper.color,
-            modelMatrix
+            modelMatrix,
           );
         }
       }