+++ systems/renderer/skybox.js
@@ -38,7 +38,7 @@ export default ({ ctx, resourceCache }) => {
         console.warn(
           NAMESPACE,
           this.type,
-          `reflectionProbe component missing _reflectionProbe. Add a reflectionProbeSystem.update(entities, { renderers: [skyboxRendererSystem] }).`
+          `reflectionProbe component missing _reflectionProbe. Add a reflectionProbeSystem.update(entities, { renderers: [skyboxRendererSystem] }).`,
         );
       } else {
         return true;
@@ -49,7 +49,7 @@ export default ({ ctx, resourceCache }) => {
         console.warn(
           NAMESPACE,
           this.type,
-          `skybox component missing texture. Provide a "envMap" or add a skyboxSystem.update(world.entities).`
+          `skybox component missing texture. Provide a "envMap" or add a skyboxSystem.update(world.entities).`,
         );
       } else {
         return true;
@@ -138,7 +138,7 @@ export default ({ ctx, resourceCache }) => {
               backgroundBlur: entity.skybox.backgroundBlur,
               outputEncoding: renderView.outputEncoding || ctx.Encoding.Linear,
               reflectionProbeEntity: entities.find(
-                (entity) => entity.reflectionProbe
+                (entity) => entity.reflectionProbe,
               ),
             });
             // entity._skybox.draw(renderView.camera, {